move RedirectorTransformer under loading/shared package
