+++ src/main/java/com/gtnewhorizons/angelica/loading/AngelicaTweaker.java
@@ -42,7 +42,7 @@ import java.util.Set;
 //@IFMLLoadingPlugin.SortingIndex(Integer.MAX_VALUE - 5)
 @IFMLLoadingPlugin.MCVersion("1.7.10")
 @IFMLLoadingPlugin.TransformerExclusions({
-    "com.gtnewhorizons.angelica.transform.RedirectorTransformer",
+    "com.gtnewhorizons.angelica.loading",
     "com.gtnewhorizons.angelica.glsm.GLStateManager"})
 public class AngelicaTweaker implements IFMLLoadingPlugin, IEarlyMixinLoader {
