+++ src/main/java/com/gtnewhorizons/angelica/loading/fml/transformers/SodiumBlockTransformer.java
@@ -2,7 +2,7 @@ package com.gtnewhorizons.angelica.loading.fml.transformers;
 
 import com.google.common.collect.ImmutableList;
 import com.gtnewhorizons.angelica.loading.shared.AngelicaClassDump;
-import com.gtnewhorizons.angelica.transform.RedirectorTransformer;
+import com.gtnewhorizons.angelica.loading.shared.RedirectorTransformer;
 import net.minecraft.launchwrapper.IClassTransformer;
 import org.apache.commons.lang3.tuple.Pair;
 import org.objectweb.asm.ClassReader;
