+++ src/main/resources/META-INF/rfb-plugin/angelica.properties
@@ -2,7 +2,7 @@ name=Angelica
 version=${version}
 additionalVersions=
 className=com.gtnewhorizons.angelica.loading.rfb.AngelicaRfbPlugin
-transformerExclusions=com.gtnewhorizons.angelica.transform.RedirectorTransformer
+transformerExclusions=com.gtnewhorizons.angelica.loading.shared
 versionConstraints=
 loadBefore=
 loadAfter=mixin; lwjgl3ify
